+++ software/main_hcs.py
@@ -19,9 +19,12 @@ if __name__ == "__main__":
     else:
         c=gui.HCSController()
         c.acquire(
-            [(1,1)],
-            ["Fluorescence 561 nm Ex"],
-            "/home/pharmbio/Downloads/testdirfordata",
+            well_list=[(1,1)], # (0,0) is A1
+            channels=["Fluorescence 561 nm Ex"],
+            experiment_id="/home/pharmbio/Downloads/testdirfordata",
+            # grid_data
+            # af_channel
+            # plate_type
         ).finished.connect(
             lambda:c.acquire(
                 [(2,2)],
